remove empty data page
