--- centreon/www/front_src/src/CloudNotificationsConfiguration/ListingPage/index.tsx
@@ -1,3 +1,0 @@
-import ListingPage from './ListingPage';
-
-export default ListingPage;
